separate all the components in different files
